+++ src/test/java/domain/ParticipantTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import domain.participant.Participant;
+import domain.user.Participant;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
