+++ src/test/java/domain/ParticipantsTest.java
@@ -1,6 +1,6 @@
 package domain;
 
-import domain.participant.Participant;
+import domain.user.Participant;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.InstanceOfAssertFactories;
